Allow CLI to display version with `copier --version`
This is a regression from #162 where `__version__` was removed due to conflicts with how black and poetry-dynamic-versioning work.

Now those conflicts are fixed and `copier --version` works again. This was what it was returning before:

```bash
➤ copier --version
copier (version not set)
```

BEWARE: In a development, it will always return `0.0.0`, but this is expected because poetry-dynamic-versioning autopatches this on release.

In `.pre-commit-config.yaml`, `require_serial: true` is added to all local tasks that run under `poetry run`. This is because when entering `poetry run`, poetry-dynamic-versioning does its magic and patches the versions in `pyproject.toml` and `__init__.py`, and unpatches them when ending the execution. If running in parallel (default pre-commit behavior), some patches are reverted and some aren't, producing random failed lints.

@Tecnativa TT20357