Fix updates getting diffed without proper answers
There was a bug in how update diff was computed. As you can see, the temporary copy that was made to produce that diff didn't get the same answers as the current project. Thus, every time the user answered something different, a diff was generated and a `.rej` file was created.

Just try this and you'll see:

```bash
copier -f update
git add .
git commit -m updated
copier -f update
git status
```

The 2nd time you update, all files should be identical and untouched. However, that wasn't the case.

I took the chance to update this test's fixture from `dst` to `tmpdir`, which is more standard and comfortable to debug.

@Tecnativa TT20357