Fix base copy answers
When somebody calls `copier update` to reapply a template, respecting subproject evolution, but just to apply some changes to some answers, the updatediff system reverted those changes.

Now they are respected, so if calling `copier update` with new data, that new data will be applied to the subproject.

The basic fix is that when creating the temporary copy in the old commit (the one used to compare), it is created using last answers obtained from the subproject, excluding any new answers passed by the user.

#169 improved the situation but certainly it didn't fix it completely. This hardened test proves it works better now.