fix: multiline default values prompted
When the question is YAML or JSON and it is multiline and has a default value, that one should be prompted in multiline mode automatically.

It includes a test that will always fail (thus it is skipped). Gotta fix that some day when I'm more pro in python-prompt-toolkit stuff. For the time being, manual test is enough and this works!

I also removed some garbage found here and there.

Fixes https://github.com/copier-org/copier/issues/298.