feat: add support to parse yaml files as data input
The --data/-d option now accepts .yml files instead of only key
value pairs. It parses the the yaml file and updates the data
dictionary with the content.

This shall make it easier for the user to configure more complex
projects. It also allows sending someone your project template
configuration.