fix: properly support diffs over updates with new interactive answers
Just like 802d0f96548c3c8a0bf1128f8f18da4cf5ca05ad, but this time we make sure that interactive answers also work as expected.

Basically, I moved the creation of the updated and clean worker after the copying of the real destination worker. With this change, we can ask the user for answers and make sure we use them when producing the clean and updated copy.

@moduon MT-1075