fix(recopy): never clone old template (even less if it's just for cleanup)
Before this patch, when doing a recopy for a subproject previously placed on a git ref that was no longer reachable, the recopy process still tried to clone the old template.

The funniest part is that it was only trying to clone it while doing the process of cleanup. Thus, even if that clone worked, it was a waste.

This should make recopy more resilient and performant.