build(deps): bump jinja
To do so, it required to add the `python_version < 4` marker to dependencies that still use upper bound markers.

I opened https://github.com/mtkennerly/dunamai/issues/74 to the only relevant one. In the case of poetry-dynamic-versioning, it's just a build-time dependency, so I don't really care it has that marker, as long as I build using Python 3.