fix(updating): never recreate deleted paths on update
Previously, paths that matched a pattern in `skip_if_exists` and that
were deleted in the generated project were recreated during each update.

These deletions are contained in the diff between old vanilla generated
project and the (pre-update) current one. They are not executed though
since the paths are excluded from the diff application.

Similarly, if a file that was deleted intentionally in a generated
project was changed in the template, an update would have regenerated
it (but not if it did not have any changes). I'm not completely sure how
this manifested, but it probably involved a suppressed merge conflict.

This commit ensures that update behavior is always consistent -
never regenerate deleted paths during `update` - by acquiring a list of
deleted paths and excluding them in the copy operations that are
relevant for acquiring the update diff.