fix(external_data): replay old external data before updating
Before this fix, when a template used local `_external_data`, it could never again be cleanly updated. It always gave conflicts.

This was happening because the old worker that did the replay always happened on an empty directory. Thus it never had external data available. Since external data can alter the render result (just like any other answer), the replay was never reliable.

Now, new data is saved in the answers file to be able to reliably replay including that context. Then, updates can work again as expected and result in meaningful diffs only when there should be.

BREAKING CHANGE: A new `_dst_commit` is added to the answers file, pointing to the commit just before the last template update. This is needed for reliable replays. This means that any copy or update operation will always produce diff when the subproject is git-tracked.