test(test_config.py): add tests for circular transclusion failure case
Circular transclusions lead to:
```bash
E           RecursionError: maximum recursion depth exceeded while calling a Python object

/home/gitpod/.local/share/uv/python/cpython-3.9.23-linux-x86_64-gnu/lib/python3.9/pathlib.py:646: RecursionError
```
It might be nice to catch as soon as it happens, and provides a more comprehensive error message including the names of the files that collide.
The tests are skipped to not intefer with normal condition testing, and commented with a TODO with a line to the one right before this one.

> NOTE: The test with `demo_transclude_circular_bis` will fail because of the circular recursion.
> The test with `demo_transclude_circular` will:
> - currently fail with a `yaml.composer.ComposerError: expected a single document in the stream`
> - if `https://github.com/copier-org/copier/pull/2251` is accepted, also fail because of the circular recursion.
> But it might be a better test as it needs to go to the deepest transclusion to get the `_excludes`.